refactor(rvr-algebra): common IR for setup instructions
