+++ extensions/algebra/rvr/src/common.rs
@@ -41,6 +41,46 @@ impl<K: ArithKind> FieldArithInstr<K> {
     }
 }
 
+/// Generic IR node for field setup (SETUP_ADDSUB, SETUP_MULDIV, SETUP_ISEQ).
+/// Use the type aliases [`crate::ModSetupInstr`] / [`crate::Fp2SetupInstr`]
+/// rather than naming this type directly.
+#[derive(Debug, Clone)]
+pub struct FieldSetupInstr {
+    c_fn: &'static str,
+    pub rd_reg: Reg,
+    pub rs1_reg: Reg,
+    pub rs2_reg: Reg,
+    pub num_limbs: u32,
+}
+
+impl FieldSetupInstr {
+    pub fn new(c_fn: &'static str, rd_reg: Reg, rs1_reg: Reg, rs2_reg: Reg, num_limbs: u32) -> Self {
+        Self { c_fn, rd_reg, rs1_reg, rs2_reg, num_limbs }
+    }
+}
+
+impl ExtInstr for FieldSetupInstr {
+    fn opname(&self) -> &str {
+        &self.c_fn["rvr_ext_".len()..]
+    }
+
+    fn emit_c(&self, ctx: &mut dyn ExtEmitCtx) {
+        let rd = ctx.read_reg(self.rd_reg);
+        let rs1 = ctx.read_reg(self.rs1_reg);
+        let rs2 = ctx.read_reg(self.rs2_reg);
+        let num_limbs = format!("{}u", self.num_limbs);
+        ctx.extern_call(self.c_fn, &["state", &rd, &rs1, &rs2, &num_limbs]);
+    }
+
+    fn clone_box(&self) -> Box<dyn ExtInstr> {
+        Box::new(self.clone())
+    }
+
+    fn is_block_end(&self) -> bool {
+        false
+    }
+}
+
 impl<K: ArithKind> ExtInstr for FieldArithInstr<K> {
     fn opname(&self) -> &str {
         K::opname()
